support github.com/ncruces/go-sqlite3 args
